Created template.py and README.md for PART III
Template
A file called 'template.py' has been created to  serve as a project template for future data science projects. It contains line and block comments that can be viewed to be mile markers or headings and sub-headings for the code of the project.

README.md
A README.md file has been created for documenting the progress of Part III.
It also contains some of the organizationsal theory behind how the projects should proceed.