Create README.md
I have created a README.md file for this project. This will serve as the primary documentation and reference as thee rpoject progresses.
The file conforms to GitHub Flavored Markdown Spec (Version 0.29-gfm).

It contains the following sections about the respository, the dataset, steps of the project, project files, future steps and images.

I hope to frequently update the README.md file throughout the project.