Data Loading
I created a file called 'iris.py' to hold all the functional code for
the project. The file contains a class 'iris' and it will be used as
the primary object of this project.

The data has been loaded from the file 'iris.data' into a pandas
dataframe in the class __init__ function.