Data Summarization
A statistical analyis of the data has been conducted and the results
summarized and printed to stdout.

To create an accurate model of the data, we must take a good look at
the data and create a statistical summary of it. We now know the
characteristics of the data and will be able to visualize it for
further understanding.