Model Selection
We have created a 10k Kfold Test harness to spot check models for their
accuracy. After evaluating the models and summarizing and visualization
of results, we find the SVM Classifier model to have the highest
accuracy score.

Spot checking models is an important step in selecting a suitable model.
It has the following benefits:
 - We find out which models perform well on our data
 - We can find out which models might suit our problem before proceeding
   with predictions
 - We can trial a number of models before committing to them

Here, we find that KNN and SVC have the highest accuracy score.