Making Predictions using Selected Model and Summarization of Results
We have fit the KNN model to the training data and made predictions
using it on the test data. After that we have scored the accuracy,
created a confusion matrix and a classification report using the
validation dataset.

To ensure that we are getting as accurate a model as possible, we test
the model using a validation dataset. After that, we can use various
tools to record the results.