Data Visualization
We have created Unimodal and Multimodal visualizations of the data.

Data visualizations help us to learn more about the data than just a
static text description. They are a powerful tool to better understand
the data in a clean and clear manner.

Through the visualizations we have learned that the data is bimodal and
highly correlated for some attributes. The data would surely benefit
from some PCA and standardizations and normalizations.