Tests for partitioning dataset
Created tests for partitioning the dataset into training and validation
datasets.

The length of the arrays created will be used to verify that the
datasets were partitioned properly. Since a 2:8 ratio is being used to
create the training and validation datasets respectively, out of 506
instances 404 will be chosen for the training dataset, and the
remaining 102 will be used to form the validation dataset.