Standardization of Dataset & Model Tuning
We have scaled the data and found KNN had the best performance.
Also, after tuning the parameters for KNN, we found  k = 3 to be the
optimal value.

The data has disparate scales, which inhibits the performance of
algorithms like KNN. Scaling the data using StandardScaler transforms
the dataset into something our algorithms can perform better in.

After finding KNN had the best performance, we sought to find the
optimal parameters for the algorithm.